backend/mouhapcap-bknd/requirements.txt
anyio==3.6.2


appdirs==1.4.4


asgiref==3.6.0

click==8.1.3


colorama==0.4.6


crispy-bootstrap5==0.7
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.13.0
django-crispy-forms==2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
dpkt==1.9.8
fastapi==0.95.1
filelock==3.12.2
h11==0.14.0
idna==3.4
lxml==4.9.3
mongoengine==0.27.0

node==1.2.1
npm==0.1.1
odict==1.9.0
optional-django==0.1.0
packaging==23.1
platformdirs==3.9.1
plumber==1.7

psycopg2-binary==2.9.7
pydantic==1.10.7
pymongo==4.5.0
pyshark==0.6
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
scapy==2.5.0
sniffio==1.3.0
sqlparse==0.2.4
starlette==0.26.1
termcolor==2.3.0
typing_extensions==4.5.0
tzdata==2023.2
uvicorn==0.21.1
virtualenv==20.24.0
zope.component==6.0
zope.deferredimport==4.4
zope.deprecation==5.0
zope.event==4.6
zope.hookable==5.4
zope.interface==6.0
zope.lifecycleevent==4.4
zope.proxy==5.0.0
